+++ source/types.ts
@@ -4,6 +4,10 @@ export type CliFlags = {
 	force: boolean;
 	workspace?: string;
 	config?: string;
+	limit?: number;
+	from?: string;
+	to?: string;
+	message?: string;
 	detach: boolean;
 	api?: string;
 	prune: boolean;